Move clone_replace tests to tests.graph.test_basic
